Extract document_config tests into document_config/tests.rs
